requirements.txt
nltk
tensorflow
djangorestframework==3.14.0
Django==4.2.1
openai